requirements.txt
annotated-types==0.5.0
anyio==3.7.1
beautifulsoup4==4.11.1
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.4
fastapi==0.100.0
h11==0.14.0
httpcore==0.17.3
httptools==0.6.0
httpx==0.24.1
idna==3.4
pydantic==2.0.2
pydantic_core==2.1.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
requests==2.28.1
sniffio==1.3.0
soupsieve==2.3.2.post1
starlette==0.27.0
typing_extensions==4.7.1
urllib3==1.26.13
uvicorn==0.22.0
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.3
